+++ ui/src/ErrorBoundary.tsx
@@ -1,7 +1,5 @@
 import { Component, FC, ReactNode, ErrorInfo } from "react";
 
-import { captureException } from "@sentry/browser";
-
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faBomb } from "@fortawesome/free-solid-svg-icons/faBomb";
 
@@ -73,9 +71,6 @@ class ErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
   componentDidCatch(error: Error, { componentStack }: ErrorInfo): void {
     if (this.state.cachedError === null) {
       this.setState({ cachedError: error });
-      captureException(error, {
-        contexts: { react: { componentStack } },
-      });
     }
     // reload after 60s, this is to fix wall monitors automatically
     // but only if the timer isn't set yet